+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prompt for DJ Controls set to receive emergencies or accountability notifications when closing.
 - Management of the airing of underwritings through DJ Controls.
 - locking of hosts to a specific DJ (or no DJs) via manage hosts. When a DJ Controls is locked to a specific DJ, it will prevent starting any shows/broadcasts that are unscheduled or not under the DJ's name, and also will prevent interfering with other shows.
+- Ability to change finalized show times in the schedule generator.
 
 ### Changed
 - Timesheet window in admin options -> manage directors -> timesheets now displays timesheets as vertical Gantt Charts.